add SGD example
